refactor folders structure
