+++ product-service/utils/productList.ts
